+++ src/ext/java/org/opentripplanner/ext/transmodelapi/TransmodelAPI.java
@@ -3,6 +3,10 @@ package org.opentripplanner.ext.transmodelapi;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import graphql.ExecutionResult;
 import graphql.schema.GraphQLSchema;
+import io.micrometer.core.instrument.Tag;
+import java.util.Collection;
+import java.util.stream.Collectors;
+import javax.ws.rs.core.HttpHeaders;
 import org.opentripplanner.api.json.GraphQLResponseSerializer;
 import org.opentripplanner.ext.transmodelapi.mapping.TransitIdMapper;
 import org.opentripplanner.ext.transmodelapi.support.GqlUtil;
@@ -42,6 +46,7 @@ public class TransmodelAPI {
 
     private static GqlUtil gqlUtil;
     private static GraphQLSchema schema;
+    private static Collection<String> tracingHeaderTags;
 
     private final Router router;
 
@@ -66,13 +71,14 @@ public class TransmodelAPI {
      * was done more explicit and enforced, not relaying on a "static" setup method to be called.
      */
     public static void setUp(
-        boolean hideFeedId,
+        TransmodelAPIParameters config,
         Graph graph,
         RoutingRequest defaultRoutingRequest
     ) {
-        if(hideFeedId) {
+        if(config.hideFeedId()) {
           TransitIdMapper.setupFixedFeedId(graph.getAgencies());
         }
+        tracingHeaderTags = config.tracingHeaderTags();
         gqlUtil = new GqlUtil(graph.getTimeZone());
         schema = TransmodelGraphQLSchema.create(defaultRoutingRequest, gqlUtil);
     }
@@ -89,7 +95,11 @@ public class TransmodelAPI {
     @POST
     @Path("/graphql")
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response getGraphQL(HashMap<String, Object> queryParameters, @HeaderParam("OTPMaxResolves") @DefaultValue("1000000") int maxResolves) {
+    public Response getGraphQL(
+            HashMap<String, Object> queryParameters,
+            @HeaderParam("OTPMaxResolves") @DefaultValue("1000000") int maxResolves,
+            @Context HttpHeaders headers
+    ) {
         if (queryParameters==null || !queryParameters.containsKey("query")) {
             LOG.debug("No query found in body");
             throw new BadRequestException("No query found in body");
@@ -110,20 +120,43 @@ public class TransmodelAPI {
         } else {
             variables = new HashMap<>();
         }
-        return index.getGraphQLResponse(query, router, variables, operationName, maxResolves);
+        return index.getGraphQLResponse(
+            query,
+            router,
+            variables,
+            operationName,
+            maxResolves,
+            getTagsFromHeaders(headers)
+        );
     }
 
     @POST
     @Path("/graphql")
     @Consumes("application/graphql")
-    public Response getGraphQL(String query, @HeaderParam("OTPMaxResolves") @DefaultValue("1000000") int maxResolves) {
-        return index.getGraphQLResponse(query, router, null, null, maxResolves);
+    public Response getGraphQL(
+            String query,
+            @HeaderParam("OTPMaxResolves") @DefaultValue("1000000") int maxResolves,
+            @Context HttpHeaders headers
+    ) {
+        return index.getGraphQLResponse(
+            query,
+            router,
+            null,
+            null,
+            maxResolves,
+            getTagsFromHeaders(headers)
+        );
     }
 
     @POST
     @Path("/graphql/batch")
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response getGraphQLBatch(List<HashMap<String, Object>> queries, @HeaderParam("OTPTimeout") @DefaultValue("10000") int timeout, @HeaderParam("OTPMaxResolves") @DefaultValue("1000000") int maxResolves) {
+    public Response getGraphQLBatch(
+            List<HashMap<String, Object>> queries,
+            @HeaderParam("OTPTimeout") @DefaultValue("10000") int timeout,
+            @HeaderParam("OTPMaxResolves") @DefaultValue("1000000") int maxResolves,
+            @Context HttpHeaders headers
+    ) {
         List<Callable<ExecutionResult>> futures = new ArrayList<>();
 
         for (Map<String, Object> query : queries) {
@@ -141,16 +174,34 @@ public class TransmodelAPI {
             }
             String operationName = (String) query.getOrDefault("operationName", null);
 
-            futures.add(() -> index.getGraphQLExecutionResult((String) query.get("query"), router,
-                    variables, operationName, maxResolves));
+            futures.add(() -> index.getGraphQLExecutionResult(
+                (String) query.get("query"),
+                router,
+                variables,
+                operationName,
+                maxResolves,
+                getTagsFromHeaders(headers)
+            ));
         }
 
         try {
             List<Future<ExecutionResult>> results = index.threadPool.invokeAll(futures);
-            return Response.status(Response.Status.OK).entity(GraphQLResponseSerializer.serializeBatch(queries, results)).build();
+            return Response
+                .status(Response.Status.OK)
+                .entity(GraphQLResponseSerializer.serializeBatch(queries, results))
+                .build();
         } catch (InterruptedException e) {
             LOG.error("Batch query interrupted", e);
             throw new RuntimeException(e);
         }
+    }
+
+    private static Iterable<Tag> getTagsFromHeaders(HttpHeaders headers) {
+        return tracingHeaderTags.stream()
+            .map(header -> {
+                String value = headers.getHeaderString(header);
+                return Tag.of(header, value == null ? "__UNKNOWN__" : value);
+            })
+            .collect(Collectors.toList());
     }
 }