+++ src/main/java/org/opentripplanner/netex/NetexBundle.java
@@ -128,6 +128,7 @@ public class NetexBundle implements Closeable {
             });
         }
         mapper.finnishUp();
+        NetexDocumentParser.finnishUp();
     }
 
     /**