+++ src/main/java/org/opentripplanner/netex/mapping/StopMapper.java
@@ -6,6 +6,7 @@ import org.opentripplanner.graph_builder.DataImportIssueStore;
 import org.opentripplanner.model.FareZone;
 import org.opentripplanner.model.Station;
 import org.opentripplanner.model.Stop;
+import org.opentripplanner.model.TransitMode;
 import org.opentripplanner.model.WgsCoordinate;
 import org.opentripplanner.netex.issues.QuayWithoutCoordinates;
 import org.opentripplanner.netex.mapping.support.FeedScopedIdFactory;
@@ -29,7 +30,12 @@ class StopMapper {
    * Map Netex Quay to OTP Stop
    */
   @Nullable
-  Stop mapQuayToStop(Quay quay, Station parentStation, Collection<FareZone> fareZones) {
+  Stop mapQuayToStop(
+          Quay quay,
+          Station parentStation,
+          Collection<FareZone> fareZones,
+          TransitMode transitMode
+  ) {
     WgsCoordinate coordinate = WgsCoordinateMapper.mapToDomain(quay.getCentroid());
 
     if (coordinate == null) {
@@ -49,7 +55,7 @@ class StopMapper {
         fareZones,
         null,
         null,
-        null
+        transitMode
     );
     stop.setParentStation(parentStation);