+++ src/main/java/org/opentripplanner/routing/algorithm/raptor/router/TransitRouter.java
@@ -114,6 +114,7 @@ public class TransitRouter {
                     requestTransitDataProvider.stopNameResolver(),
                     router.graph.getTransferService(),
                     requestTransitDataProvider,
+                    transitLayer.getStopIndex().stopBoardAlightCosts,
                     raptorRequest,
                     request.transferOptimization
             ).optimize(transitResponse.paths());