+++ src/main/java/org/opentripplanner/routing/algorithm/transferoptimization/api/TransferOptimizationParameters.java
@@ -35,4 +35,21 @@ public interface TransferOptimizationParameters {
    * </pre>
    */
   double minSafeWaitTimeFactor();
+
+  /**
+   * Use this to add an extra board- and alight-cost for (none) prioritized stops. A {@code
+   * stopBoardAlightCosts} is added to the generalized-cost during routing. But, this cost cannot
+   * be too high, because that would add extra cost to the transfer, and favor other alternative
+   * paths. But, when optimizing transfers, we do not have to take other paths into consideration
+   * and can "boost" the stop-priority-cost to allow transfers to take place at a preferred stop.
+   * <p>
+   * The cost added during routing is already added to the generalized-cost used as a base in the
+   * optimized transfer calculation. By setting this parameter to 0, no extra cost is added, by
+   * setting it to {@code 1.0} the stop-cost is doubled.
+   * <p>
+   * Stop priority is only supported by the NeTEx import, not GTFS.
+   * <p>
+   * Default value is 0.0.
+   */
+  double extraStopBoardAlightCostsFactor();
 }